add cli class
