+++ phpborg.php
@@ -1,164 +1,20 @@
 #!/usr/bin/php -q
 <?php
 
-// TODO this isn't a nice way to declare variables, I suggest using classes and constructors to set properties
+use phpBorg\Cli;
+
 $flog = fopen('/var/log/backup.log', 'a+');
 declare(ticks=1);
 pcntl_signal(SIGTERM, "signal_handler");
 pcntl_signal(SIGINT, "signal_handler");
 
-/**
- * @param $inputSeconds
- * @return string
- */
-function secondsToTime($inputSeconds)
-{
-    $secondsInAMinute = 60;
-    $secondsInAnHour = 60 * $secondsInAMinute;
-    $secondsInADay = 24 * $secondsInAnHour;
 
-    // Extract days
-    $days = floor($inputSeconds / $secondsInADay);
+require 'lib/Cli.php';
+$cli = new Cli();
 
-    // Extract hours
-    $hourSeconds = $inputSeconds % $secondsInADay;
-    $hours = floor($hourSeconds / $secondsInAnHour);
 
-    // Extract minutes
-    $minuteSeconds = $hourSeconds % $secondsInAnHour;
-    $minutes = floor($minuteSeconds / $secondsInAMinute);
 
-    // Extract the remaining seconds
-    $remainingSeconds = $minuteSeconds % $secondsInAMinute;
-    $seconds = ceil($remainingSeconds);
 
-    // Format and return
-    $timeParts = [];
-    $sections = [
-        'd' => (int)$days,
-        'h' => (int)$hours,
-        'm' => (int)$minutes,
-        's' => (int)$seconds,
-    ];
-
-    foreach ($sections as $name => $value) {
-        if ($value > 0) {
-            $timeParts[] = $value . '' . $name . ($value == 1 ? '' : '');
-        }
-    }
-
-    return implode('', $timeParts);
-}
-
-/**
- * @param $signal
- */
-function signal_handler($signal)
-{
-    switch ($signal) {
-        case SIGTERM:
-        case SIGINT:
-            fwrite($flog, "\nControlC detected => Script terminating\n");
-            exit;
-    }
-}
-
-/**
- * @param $repo
- * @return array|bool|int
- */
-function parse($repo)
-{
-    if (file_exists("$repo/conf/borg.conf")) {
-
-        $exclude = $backup = "";
-        $config = parse_ini_file("$repo/conf/borg.conf");
-        $ex = explode(',', $config['exclude']);
-        foreach ($ex as $zz) {
-            $exclude .= "--exclude " . trim($zz) . " ";
-        }
-        $config['exclude'] = $exclude;
-        $back = explode(',', $config['backup']);
-        foreach ($back as $yy) {
-            $backup .= trim($yy) . " ";
-        }
-        $config['backup'] = $backup;
-        return $config;
-    } else {
-        return 0;
-    }
-}
-
-/**
- * @param $cmd
- * @param string $input
- * @return array
- */
-function my_exec($cmd, $input = '')
-{
-    $proc = proc_open($cmd, array(0 => array('pipe', 'r'), 1 => array('pipe', 'w'), 2 => array('pipe', 'w')), $pipes);
-    fwrite($pipes[0], $input);
-    fclose($pipes[0]);
-    $stdout = stream_get_contents($pipes[1]);
-    fclose($pipes[1]);
-    $stderr = stream_get_contents($pipes[2]);
-    fclose($pipes[2]);
-    $rtn = proc_close($proc);
-    return array('stdout' => $stdout,
-        'stderr' => $stderr,
-        'return' => $rtn
-    );
-}
-
-/**
- * @param $keepday
- * @param $srv
- * @return array
- */
-function prune($keepday, $srv)
-{
-    $deleted = NULL;
-    $keepday = $keepday - 1;
-    $e = my_exec("/usr/bin/borg prune --save-space --force --list --keep-daily $keepday /data0/backup/$srv/backup");
-    return ['return' => $e['return'], 'stdout' => $e['stdout'], 'stderr' => $e['stderr']];
-}
-
-/**
- * @param $file
- * @return array
- */
-function parselog($file)
-{
-    $e = my_exec("/usr/bin/borg info $file --json");
-    $json = $e['stdout'];
-    if ($e['return'] == 0) {
-        $var = json_decode($json, true);
-        $cmdline = "";
-        $commandline = $var['archives'][0]['command_line'];
-        $duration = $var['archives'][0]['duration'];
-        $start = $var['archives'][0]['start'];
-        $end = $var['archives'][0]['end'];
-        $id = $var['archives'][0]['id'];
-        $csize = $var['archives'][0]['stats']['compressed_size'];
-        $dsize = $var['archives'][0]['stats']['deduplicated_size'];
-        $nbfile = $var['archives'][0]['stats']['nfiles'];
-        $osize = $var['archives'][0]['stats']['original_size'];
-        $total_chunks = $var['cache']['stats']['total_chunks'];
-        $total_csize = $var['cache']['stats']['total_csize'];
-        $total_size = $var['cache']['stats']['total_size'];
-        $total_unique_chunks = $var['cache']['stats']['total_unique_chunks'];
-        $unique_csize = $var['cache']['stats']['unique_csize'];
-        $unique_size = $var['cache']['stats']['unique_size'];
-        $encryption = $var['encryption']['mode'];
-        $location = $var['repository']['location'];
-        $last_modified = $var['repository']['last_modified'];
-
-        foreach ($commandline as $value) {
-            $cmdline .= "$value ";
-        }
-        return array('archive_id' => $id, 'dur' => $duration, 'start' => $start, 'end' => $end, 'csize' => $csize, 'dsize' => $dsize, 'nfiles' => $nbfile, 'osize' => $osize, 'ttchunks' => $total_chunks, 'ttcsize' => $total_csize, 'ttsize' => $total_size, 'ttuchunks' => $total_unique_chunks, 'ucsize' => $unique_csize, 'usize' => $unique_size, 'location' => $location, 'modified' => $last_modified, 'encryption' => $encryption, 'error' => NULL);
-    } else return array('error' => 1, 'stderr' => $e['stderr'], 'stdout' => $e['stdout']);
-}
 
 /**
  * @param $server
@@ -172,7 +28,7 @@ function db_connect($server)
             return $link;
             break;
     }
-
+	return false;
 }
 
 /**
@@ -218,7 +74,7 @@ function fsql_row($reqsql, $db)
 /**
  * @param $path
  */
-function makestat($path)
+function makestat($path, $cli)
 {
     $db = db_connect('extranet');
     $tmp = NULL;
@@ -229,7 +85,7 @@ function makestat($path)
         foreach ($backupfile as $value) {
             $file = $dir . "/backup::" . $value['archive'];
             $repo = $dir . "/backup";
-            $info = parselog("$file");
+            $info = $cli->parselog("$file");
             $insert = "INSERT IGNORE INTO archives VALUE (NULL,'$repo','$value[archive]','$info[archive_id]','$info[dur]','$info[start]','$info[end]','$info[csize]','$info[dsize]','$info[osize]','$info[nfiles]')";
             $update = "UPDATE repository set `size` = '$info[ttsize]', `dsize` = '$info[ucsize]',`csize`= '$info[ttcsize]', `ttuchunks` = '$info[ttuchunks]', `ttchunks` = '$info[ttchunks]' where nom = '$repo'";
             mysqli_query($db, $insert) or $tmp .= mysqli_error($db);
@@ -248,7 +104,7 @@ if (!empty($argv[1])) {
 }
 
 if ($param == "prune") {
-    $result = prune(3, "yodaa");
+    $result = $cli->prune(3, "yodaa");
     echo "RETURN = $result[return]\n";
     if ($result['return'] == 0) {
         $separator = "\r\n";
@@ -271,7 +127,7 @@ if ($param == "prune") {
 if ($param == "archiveinfo") {
     $path = "/data0/backup/*";
     if ($argv[2]) $path = "/data0/backup/$argv[2]";
-    makestat($path);
+    makestat($path, $cli);
 }
 
 
@@ -284,7 +140,7 @@ if ($param == "backup") {
     }
 
     $db = db_connect('extranet');
-    $config = parse("/data0/backup/$srv");
+    $config = $cli->parse("/data0/backup/$srv");
 
     if (!$config) die("Error, the server does not exist or the conf is not good\n");
     fsql("INSERT INTO `report` (`start`) VALUES (Now())", $db, "0");
@@ -302,7 +158,7 @@ if ($param == "prune") {
     $error = $nb_archive = $tmplog = $dur = $osize = $csize = $dsize = $nfiles = NULL;
 
     foreach (glob('/data0/backup/*', GLOB_ONLYDIR) as $dir) {
-        $config = parse($dir);
+        $config = $cli->parse($dir);
         echo "Serveur $config[host]\n";
 
         if (!$config) {
@@ -349,7 +205,7 @@ if ($param == "full") {
         } else {
             mysqli_query($db, "UPDATE IGNORE report  set `curpos`= '$config[host]' WHERE id=$idlog");
             fwrite($flog, "Verification of the rules of conservation\n");
-            $result = prune($config['retention'], $config['host']);
+            $result = $cli->prune($config['retention'], $config['host']);
             if ($result['return'] == 0) {
                 $separator = "\r\n";
                 $line = strtok($result['stderr'], $separator);
@@ -371,11 +227,11 @@ if ($param == "full") {
             system("/bin/chown -R $config[host]:$config[host] $config[repo]");
             system("/bin/chmod -R 700 $config[repo]");
             fwrite($flog, "Execution of the backup\n");
-            $e = my_exec("ssh -p $config[port] -tt $config[host] \"/usr/bin/borg create  --compression $config[compression] $config[exclude] ssh://$config[host]@10.10.69.15$config[repo]::backup_$date $config[backup]\"");
+            $e = $cli->my_exec("ssh -p $config[port] -tt $config[host] \"/usr/bin/borg create  --compression $config[compression] $config[exclude] ssh://$config[host]@10.10.69.15$config[repo]::backup_$date $config[backup]\"");
             if ($e['return'] == '0') {
-                $info = parselog("$config[repo]::backup_$date");
+                $info = $cli->parselog("$config[repo]::backup_$date");
                 if (!$info['error']) {
-                    $durx = secondsToTime($info['dur']);
+                    $durx = $cli->secondsToTime($info['dur']);
                     fwrite($flog, "Backup terminé en $durx secondes\n");
                     $archname = "backup_$date";
                     $dur += $info['dur'];
@@ -424,6 +280,6 @@ if ($param == "info") {
         exit(1);
     }
 
-    $log = parselog($file);
+    $log = $cli->parselog($file);
     print_r($log);
 }
